requirements.txt
setuptools
binance-futures-connector>=4.0.0
numpy==1.26.4
pycoingecko==3.1.0
requests
pytz
pandas[excel]
lxml
peewee
urllib3>=1.26.9,<2.0.0
openai
fpdf2==2.7.9
pycodestyle
pytest
pytest-cov
TA-Lib